app/requirements.txt
alembic
bcrypt==4.0.1
cloudinary
fastapi
fastapi_jwt_auth
fastapi_mail
google-cloud-storage
jinja2
inflect
pandas
passlib[bcrypt]
pathlib
pillow
psycopg2-binary
pydantic
pydantic-settings
python-dateutil
#python-dotenv
python-multipart
python-jose[cryptography]
qrcode
starlette
sqlalchemy
uuid
uvicorn
cloud-sql-python-connector==1.2.4
functions-framework==3.3.0
pg8000==1.29.8
python-Levenshtein
fastapi_limiter
slowapi
certbot
asyncpg
httpx